TestRunGameTask: split junit4 vintage tests
so other tests can use jupiter Extension
